+++ analyzers/__init__.py
@@ -2,5 +2,6 @@
 
 from .voc_analyzer import VOCAnalyzer
 from .priority_ranker import PriorityRanker
+from .decision_analyzer import DecisionAnalyzer
 
-__all__ = ["VOCAnalyzer", "PriorityRanker"]
+__all__ = ["VOCAnalyzer", "PriorityRanker", "DecisionAnalyzer"]